+++ apps/dashboard/app/javascript/utils.js
@@ -1,5 +1,4 @@
 import {analyticsPath} from "./config";
-import { ariaNotify } from './notify';
 
 export function cssBadgeForState(state){
   switch (state) {
@@ -98,6 +97,30 @@ export function openLinkInJs(event) {
     const alertDiv = document.createElement('div');
     alertDiv.innerHTML = html.split("ALERT_MSG").join(msg);
     mainDiv.prepend(alertDiv);
+  }
+}
+
+// Sets a custom message for a generic ARIA live region
+export function ariaNotify(message) {
+  const liveRegion = document.getElementById("aria_live_region");
+
+  if(liveRegion) {
+    liveRegion.textContent = message;
+  }
+}
+
+// Push a notification to the user using the Notification API
+export function pushNotify(message, options = {}) {
+  if (!("Notification" in window)) return;
+
+  if (Notification.permission === "granted") {
+    new Notification(message, options);
+  } else if (Notification.permission !== "denied") {
+    Notification.requestPermission().then(permission => {
+      if (permission === "granted") {
+        new Notification(message, options);
+      }
+    });
   }
 }
